src/grpc/src/requirements.txt
accelerate
grpcio-reflection
grpcio-tools
grpcio-status
googleapis-common-protos
requests
pillow
python-json-logger
torch==2.*
transformers==4.*
